kernel/binary_manager: Support kernel binary update in flat build
Support binary update in flat build for kernel binary.
