+++ os/kernel/binary_manager/binary_manager.h
@@ -280,6 +280,8 @@ int binary_manager_scan_bootparam(binmgr_bpinfo_t *bp_info);
 binmgr_bpdata_t *binary_manager_get_bpdata(void);
 int binary_manager_get_inactive_path(int requester_pid, char *bin_name);
 void binary_manager_update_bootparam(int requester_pid, uint8_t type);
+void binary_manager_reset_board(int reboot_reason);
+int binary_manager_update_kernel_binary(void);
 
 /****************************************************************************
  * Binary Manager Main Thread
